requirements.txt
bitstring==3.1.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pefile==2019.4.18
#pehash==0.9.0
pycparser==2.20
pyimpfuzzy==0.5
pymongo==3.11.0
requests==2.24.0
six==1.15.0
ssdeep==3.4
urllib3==1.25.10
virustotal-api==1.1.11
Werkzeug==1.0.1
